Checkpoint Fixing some merge nits, Refactoring the string provider code, Implementing fallback mechanism
