+++ src/common/stringsProvider.ts
@@ -1,0 +1,65 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+export default class StringsProvider {
+    private _initialized: boolean = false;
+
+    public static dtWindow: any;
+    private static _instance: StringsProvider;
+    private _fallBackMap: Map<string, Array<string>> = new Map<string, Array<string>>();
+
+    private constructor() {
+        // private initialization for singleton.
+        this.fallBackInitialization();
+    }
+
+    public getStringsCallback(message: any): void {
+        if (!message)
+            return;
+
+        if (!this._initialized) {
+            // initialize
+            let data = JSON.parse(message.data);
+            StringsProvider.dtWindow.loadTimeData = { data };
+            this._initialized = true;
+        }
+
+        return;
+    }
+
+    public static get instance() {
+        if (!StringsProvider._instance)
+            StringsProvider._instance = new StringsProvider();
+
+        return StringsProvider._instance;
+    }
+
+    public static getFallback(locale: string | undefined): string {
+        if (locale && StringsProvider.instance) {
+            let fallbackLocale: string[] | undefined = StringsProvider.instance._fallBackMap.get(locale);
+
+            // the provided locale is not supported
+            if (!fallbackLocale)
+                return 'en-us';
+
+            if (fallbackLocale.length === 0) {
+                // this is a "leaf node" so a locale is assumed to 
+                // be provided by developer
+                return locale;
+            } else {
+                return StringsProvider.getFallback(fallbackLocale.pop())
+            }
+        }
+
+        // not even the last locale worked so we default to en-us
+        return locale ? locale : 'en-us';
+    }
+
+    private fallBackInitialization() {
+        // implement custom fallback logic here as follows:
+        // this._fallBackMap.set('en-us', []): if a language is available (json provided)
+        // this._fallBackMap.set('test', ['es-mx']): fallback list for test language
+        this._fallBackMap.set('en-us', ['qps-ploc'])
+        this._fallBackMap.set('qps-ploc', [])
+    }
+}
