+++ src/utils.ts
@@ -9,6 +9,7 @@ import * as os from "os";
 import * as path from "path";
 import * as url from "url";
 import * as vscode from "vscode";
+import StringsProvider from "./common/stringsProvider"
 import TelemetryReporter from "vscode-extension-telemetry";
 import packageJson from "../package.json";
 import DebugTelemetryReporter from "./debugTelemetryReporter";
@@ -361,7 +362,7 @@ export function getRuntimeConfig(config: Partial<IUserConfig> = {}): IRuntimeCon
     };
 }
 
-/**
+ /**
  * Gets the localized resources that will be displayed in the frontend
  * @param extensionPath The root folder for the extension path.
  */
@@ -369,13 +370,15 @@ export async function getLocalizedStrings(extensionPath: string): Promise<string
     try {
         let locale = "en-us";
         if (process.env.VSCODE_NLS_CONFIG) {
-            locale = JSON.parse(process.env.VSCODE_NLS_CONFIG).locale;
+            locale = StringsProvider.getFallback(JSON.parse(process.env.VSCODE_NLS_CONFIG).locale);
         }
 
-        let resourcePath = vscode.Uri.file(path.join(extensionPath, "resources", "locales", `${locale}.json`));
-        let frontendStringDocument = await vscode.workspace.openTextDocument(resourcePath);
-        if (frontendStringDocument)
-            return frontendStringDocument.getText();
+        if (locale !== "en-us") {
+            let resourcePath = vscode.Uri.file(path.join(extensionPath, "resources", "locales", `${locale}.json`));
+            let frontendStringDocument = await vscode.workspace.openTextDocument(resourcePath);
+            if (frontendStringDocument)
+                return frontendStringDocument.getText();
+        }
 
         return "";
     }
